drivers: pwm: core: Add offset support
Some hardware supports delaying the start of a PWM signal. This commit
adds support for optional PWM offset(delay) that will be applied to a
specific channel.

In a case where a pwm chip needs two or more outputs to respect a
certain timing, the offset parameter will come in handy.

This offset can be seen as a phase shift or a time delay, relative to
the first PWM pulse.

Signed-off-by: Sergiu Cuciurean <sergiu.cuciurean@analog.com>